chore(buildtool): Refactor buildtool util and subprocess forking helper functions
