+++ unittest/buildtool/util_test.py
@@ -21,7 +21,10 @@ import shutil
 import tempfile
 import unittest
 
-import buildtool.util
+from buildtool import (
+    ensure_dir_exists,
+    timedelta_string,
+    write_to_path)
 
 
 class TestRunner(unittest.TestCase):
@@ -33,91 +36,20 @@ class TestRunner(unittest.TestCase):
   def tearDownClass(cls):
     shutil.rmtree(cls.base_temp_dir)
 
-  def do_run_subprocess_ok(self, check):
-    if os.path.exists('/bin/true'):
-      true_path = '/bin/true'
-    elif os.path.exists('/usr/bin/true'):
-      true_path = '/usr/bin/true'
-    else:
-      raise NotImplementedError('Unsupported test platform.')
-
-    tests = [(true_path, ''),
-             ('/bin/echo Hello', 'Hello'),
-             ('/bin/echo "Hello"', 'Hello'),
-             ('/bin/echo "Hello World"', 'Hello World'),
-             ('/bin/echo "Hello\nWorld"', 'Hello\nWorld'),
-             ('/bin/echo \'"Hello World"\'', '"Hello World"')]
-    for cmd, expect in tests:
-      if check:
-        output = buildtool.util.check_subprocess(cmd)
-      else:
-        code, output = buildtool.util.run_subprocess(cmd)
-        self.assertEquals(0, code)
-
-      self.assertEquals(expect, output)
-
-  def test_run_subprocess_ok(self):
-    self.do_run_subprocess_ok(False)
-
-  def test_check_subprocess_ok(self):
-    self.do_run_subprocess_ok(False)
-
-  def test_run_subprocess_fail(self):
-    if os.path.exists('/bin/false'):
-      false_path = '/bin/false'
-    elif os.path.exists('/usr/bin/false'):
-      false_path = '/usr/bin/false'
-    else:
-      raise NotImplementedError('Unsupported test platform.')
-
-    got, output = buildtool.util.run_subprocess(false_path)
-    self.assertEquals((1, ''), (got, output))
-
-    got, output = buildtool.util.run_subprocess('/bin/ls /abc/def')
-    self.assertNotEquals(0, got)
-    self.assertTrue(output.find('No such file or directory') >= 0)
-
-  def test_check_subprocess_fail(self):
-    if os.path.exists('/bin/false'):
-      false_path = '/bin/false'
-    elif os.path.exists('/usr/bin/false'):
-      false_path = '/usr/bin/false'
-    else:
-      raise NotImplementedError('Unsupported test platform.')
-
-    tests = [false_path, '/bin/ls /abc/def']
-    for test in tests:
-      with self.assertRaises(Exception) as ex:
-        buildtool.util.check_subprocess(test)
-      self.assertTrue(hasattr(ex.exception, 'logged'))
-
-  def test_run_subprocess_get_pid(self):
-    # See if we can run a job by looking up our job
-    # This is also testing parsing command lines.
-    code, output = buildtool.util.run_subprocess('/bin/ps -f')
-    self.assertEquals(0, code)
-    my_pid = ' %d ' % os.getpid()
-    candidates = [line for line in output.split('\n')
-                  if line.find(my_pid) > 0 and line.find('/bin/ps') < 0]
-    if len(candidates) != 1:
-      logging.error('Unexpected output\n%s', output)
-    self.assertEquals(1, len(candidates))
-    self.assertTrue(candidates[0].find(' python ') > 0)
-
   def test_ensure_dir(self):
     want = os.path.join(self.base_temp_dir, 'ensure', 'a', 'b', 'c')
     self.assertFalse(os.path.exists(want))
-    buildtool.util.ensure_dir_exists(want)
+    ensure_dir_exists(want)
     self.assertTrue(os.path.exists(want))
 
     # Ok if already exists
-    buildtool.util.ensure_dir_exists(want)
+    ensure_dir_exists(want)
     self.assertTrue(os.path.exists(want))
 
   def test_write_to_path(self):
     path = os.path.join(self.base_temp_dir, 'test_write', 'file')
     content = 'First Line\nSecond Line'
-    buildtool.util.write_to_path(content, path)
+    write_to_path(content, path)
     with open(path, 'r') as f:
       self.assertEquals(content, f.read())
 
@@ -132,7 +64,7 @@ class TestRunner(unittest.TestCase):
         (timedelta(0, 2, 123456), '2.123 secs')
     ]
     for test in tests:
-      self.assertEquals(test[1], buildtool.util.timedelta_string(test[0]))
+      self.assertEquals(test[1], timedelta_string(test[0]))
 
 
 if __name__ == '__main__':
